ref: execute_swap
