+++ contracts/hub/stable_vlp/src/reply.rs
@@ -1,6 +1,6 @@
 use cosmwasm_std::{from_json, to_json_binary, DepsMut, Reply, Response, SubMsgResult};
 use cw_utils::parse_reply_execute_data;
-use euclid::{error::ContractError, msgs::stable_vlp::VlpSwapResponse};
+use euclid::{error::ContractError, pool::VlpSwapResponse};
 
 pub fn on_next_swap_reply(_deps: DepsMut, msg: Reply) -> Result<Response, ContractError> {
     match msg.result.clone() {
