+++ contracts/hub/vlp/src/query.rs
@@ -10,8 +10,8 @@ use euclid::token::{Pair, PairWithAmount, Token};
 
 use crate::state::{BALANCES, CHAIN_LP_TOKENS, STATE};
 use euclid::msgs::vlp::{
-    AllPoolsResponse, FeeResponse, GetLiquidityResponse, GetStateResponse, GetSwapResponse,
-    PoolInfo, PoolResponse, TotalFeesPerDenomResponse, TotalFeesResponse,
+    calculate_swap, AllPoolsResponse, FeeResponse, GetLiquidityResponse, GetStateResponse,
+    GetSwapResponse, PoolInfo, PoolResponse, TotalFeesPerDenomResponse, TotalFeesResponse,
 };
 use euclid::pool::State;
 
@@ -176,23 +176,6 @@ fn get_pool(
             .unwrap_or(Uint128::zero()),
         lp_shares: chain_lp_tokens,
     })
-}
-// Function to calculate the asset to be recieved after a swap
-pub fn calculate_swap(
-    swap_amount: Uint128,
-    reserve_in: Uint128,
-    reserve_out: Uint128,
-) -> Result<Uint128, ContractError> {
-    // Calculate the k constant product
-    let k = reserve_in.checked_mul(reserve_out)?;
-    // Calculate the new reserve of token 1
-    let new_reserve_in = reserve_in.checked_add(swap_amount)?;
-    // Calculate the new reserve of token 2
-    let new_reserve_out = k.checked_div(new_reserve_in)?;
-    // Calculate the amount of token 2 to be recieved
-    let token_2_recieved = reserve_out.checked_sub(new_reserve_out)?;
-
-    Ok(token_2_recieved)
 }
 
 pub fn calculate_lp_allocation(
